fix: keep failed followers from reshaping task creation receipts
